pyproject.toml
[project]
name = "llmamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "marimo>=0.13.11",
    "pandas>=2.2.3",
    "pydantic-ai>=0.2.9",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
]
